:recycle: Refactor : 소셜 로그인 관련 폴더 구조 정리 (#8 , #9)
