+++ src/main/java/forus/naviforyou/domain/member/dto/google/GoogleResInfo.java
@@ -1,0 +1,17 @@
+package forus.naviforyou.domain.member.dto.google;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.*;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown=true)
+public class GoogleResInfo {
+    private String id;
+    private String email;
+    private String verified_email;
+    private String name;
+    private String given_name;
+    private String family_name;
+    private String picture;
+    private String locale;
+}
