--- src/main/java/forus/naviforyou/domain/socialLogin/google/dto/googleInfoDto.java
@@ -1,6 +1,0 @@
-package forus.naviforyou.domain.socialLogin.google.dto;
-
-public class googleInfoDto {
-    private String provider;
-    private String providerId;
-}
