--- src/main/java/forus/naviforyou/domain/socialLogin/naver/repository/MemberRepository.java
@@ -1,13 +1,0 @@
-//package forus.naviforyou.domain.socialLogin.naver.repository;
-//
-//import forus.naviforyou.global.common.collection.member.Member;
-//import org.springframework.data.mongodb.repository.MongoRepository;
-//import org.springframework.stereotype.Repository;
-//
-//import java.util.Optional;
-//
-//@Repository
-//public interface MemberRepository extends MongoRepository<Member,String> {
-//    Optional<Member> findByEmail(String email);
-//
-//}
