--- src/main/java/forus/naviforyou/domain/socialLogin/naver/service/NaverUserService.java
@@ -1,5 +1,0 @@
-package forus.naviforyou.domain.socialLogin.naver.service;
-
-public interface NaverUserService {
-    String getAccessToken(String authorizationCode, String state);
-}
